Refactored/Organized controller structure
