+++ src/com/iteration3/controller/Controllers/WonderController.java
@@ -1,5 +1,7 @@
-package com.iteration3.controller;
+package com.iteration3.controller.Controllers;
 
+import com.iteration3.controller.Action;
+import com.iteration3.controller.Observer;
 import com.iteration3.model.GameModel;
 import com.iteration3.view.GameWindow;
 import com.iteration3.view.MainView;
@@ -10,39 +12,49 @@ import javafx.scene.input.KeyCode;
 import java.util.ArrayList;
 import java.util.HashMap;
 
-public class ProductionController implements Observable{
+/**
+ * Created by Clay on 4/14/2017.
+ */
+public class WonderController{
 
-    GameModel model;
-    GameWindow window;
+    private GameModel model;
+    private GameWindow window;
     ArrayList<Observer> subscribers;
     boolean lastPlayer;
     HashMap<KeyCode, Action> keyMap;
-    private EventHandler<ActionEvent> endTurn, useFactory;
-    // Location current;
-    //TODO: Iterator of transporters (or locations)
+    private EventHandler<ActionEvent> endTurn;
+    int numBricks;
 
-    public ProductionController(GameModel model, GameWindow window, HashMap<KeyCode, Action> keyMap) {
+    public WonderController(GameModel model, GameWindow window, HashMap<KeyCode, Action> keyMap) {
         this.model = model;
         this.window = window;
         this.keyMap = keyMap;
-        lastPlayer = false;
         subscribers = new ArrayList<>();
+        lastPlayer = false;
+        numBricks = 0;
 
         initializeKeyMap();
         createHandlers();
     }
 
     private void initializeKeyMap() {
-        keyMap.put(KeyCode.RIGHT, new Action() {
+        keyMap.put(KeyCode.UP, new Action() {
             public void execute() {
-                //iter.next();
+                numBricks++;
 
             }
         });
 
-        keyMap.put(KeyCode.LEFT, new Action() {
+        keyMap.put(KeyCode.DOWN, new Action() {
             public void execute() {
-                //iter.prev();
+                numBricks--;
+
+            }
+        });
+
+        keyMap.put(KeyCode.ENTER, new Action() {
+            public void execute() {
+                //model.buildBricks(numBricks);
 
             }
         });
@@ -50,21 +62,11 @@ public class ProductionController implements Observable{
     }
 
     private void createHandlers() {
-        useFactory = new EventHandler<ActionEvent>() {
-            public void handle(ActionEvent e) {
-                //model.nextPlayer();
-                lastPlayer = !lastPlayer;
-                if (lastPlayer);
-                notifyAllObservers();
-            }
-        };
-
         endTurn = new EventHandler<ActionEvent>() {
+            @Override
             public void handle(ActionEvent e) {
+
                 //model.nextPlayer();
-                lastPlayer = !lastPlayer;
-                if (lastPlayer);
-                    notifyAllObservers();
             }
         };
     }
@@ -82,6 +84,3 @@ public class ProductionController implements Observable{
             o.update();
     }
 }
-
-
-
