+++ src/com/iteration3/controller/Modes/PhaseMode.java
@@ -1,4 +1,4 @@
-package com.iteration3.controller;
+package com.iteration3.controller.Modes;
 
 /**
  * Created by Clay on 4/16/2017.
