+++ src/com/iteration3/controller/Modes/RoadMode.java
@@ -1,13 +1,14 @@
-package com.iteration3.controller;
+package com.iteration3.controller.Modes;
 
 /**
  * Created by Clay on 4/16/2017.
  */
+import com.iteration3.controller.Action;
+import com.iteration3.controller.Modes.DirectionalMode;
 import com.iteration3.model.GameModel;
 import com.iteration3.view.GameWindow;
 import javafx.scene.input.KeyCode;
 
-import java.util.ArrayList;
 import java.util.HashMap;
 
 /**
