+++ src/com/iteration3/view/WelcomeViewWindow.java
@@ -1,16 +1,10 @@
 package com.iteration3.view;
 
-import com.iteration3.controller.WelcomeViewController;
+import com.iteration3.controller.Controllers.WelcomeViewController;
 import com.iteration3.utilities.Assets;
-import javafx.event.ActionEvent;
-import javafx.event.EventHandler;
 import javafx.geometry.Pos;
-import javafx.scene.Scene;
 import javafx.scene.control.Button;
-import javafx.scene.image.Image;
 import javafx.scene.image.ImageView;
-import javafx.scene.layout.BorderPane;
-import javafx.scene.layout.HBox;
 import javafx.scene.layout.StackPane;
 import javafx.scene.layout.VBox;
 import javafx.stage.Stage;
